docs(app-session): 허브 추출의 주석·타입·문서를 적대적 검증 결과대로 고친다
앞선 세 커밋(handleKeyEvent·mouse·tick 소함수 추출)을 3라운드 적대적 검증에 넘겼다.
동작 동등성은 기계적으로 확인됐다 — 헬퍼 호출을 본문으로 역-인라인한 뒤 정규화 diff한
결과 세 함수 모두 IDENTICAL, 함수 밖 파일도 IDENTICAL(blocker 0, major 0). 살아남은
것은 전부 주석·타입 표기·문서였고 이 커밋이 그것들을 고친다. **코드 동작은 바뀌지 않는다.**

**doc comment가 고아가 됐던 것을 되돌린다.** `routeActivePointerGesture`를 `mouse`
바로 앞에 삽입하면서, 원래 `pub fn mouse`에 붙어 있던 세 줄이 새 함수에 흡수됐다.
그 결과 `kind == 2 or kind == 3`만 처리하는 함수가 "4=더블클릭, 5=트리플클릭" 계약을
자기 문서 첫 문단으로 광고하고, 여섯 줄 뒤에서 스스로 "down(1)은 여기서 처리하지
않고"라고 모순되게 말했다. 반대로 ABI 진입점인 `mouse`는 doc이 0줄이었다.

**거짓 불변식을 새로 단언한 문장을 실측대로 고친다.** `runFramePreHousekeeping`의
doc이 "파일시스템·네트워크 I/O는 worker가 소유한다"고 적었는데 예외가 둘이다 —
`pollAgentKinds`는 ≈0.5s throttle로 foreground process 열거 syscall을 직접 하고 그
안에서 surface core 락까지 잡으며, `requestBranchMenu`는 `.git` 탐색과 git 실행 파일
탐색으로 `access(2)`를 친다. 같은 doc 첫 줄이 계측 순서도 거꾸로 말했다(이 함수는
계측 *전*이 아니라 `ft_start`→`ft_pre`로 측정되는 구간 자체다) — 그대로 믿고 위로
올리면 pre 항목이 항상 0으로 붕괴한다.

**`collectFindViewSpans`의 계약을 정확히 적는다.** 재검색 게이트는 출력 유무만이
아니라 `remote_find_dirty`도 본다. span 채우기는 무조건이 아니라 Find 패널이 열려
있을 때만이다. 그리고 이 함수는 활성 surface가 로컬이면 `lockCore`를 잡는데 doc에
그 말이 없었다 — `core_mutex`는 비재진입이라 락을 쥔 자리에서 부르면 즉시 panic한다.
인자 타입도 `u64` → `usize`로 고친다(단일 출처 `app.RuntimePumpDrainSummary`가 usize).

**이동한 주석의 상대 지시어를 이름으로 바꾼다.** "아래 build/투영", "아래 일반 처리",
"tick은 그 캐시를…"이 이제 함수 밖을 가리킨다. 각각 호출부(`tick`), 호출자(`mouse`),
"이 함수"로 고쳤다. `keyConsumedByApp`의 doc은 소비 주체를 앱 UI로만 열거했는데 실제
호출처 하나는 스크롤백 페이지 스크롤(터미널 자신이 소비)이라, 열거 대신 "PTY write
없이 여기서 종결되는 모든 경로"라는 판정 기준으로 다시 썼다.

**정지 사유를 정정한다.** `cell_colors`를 안 뺀 이유로 락 보유 구간을 들었으나
검증에서 반증됐다 — 그 블록의 락은 `if` 블록 안에 갇혀 있고 `unlockCore`가 블록
마지막 문장이라 `defer`와 등가이며, 같은 작업에서 뺀 `collectFindViewSpans`가 정확히
그 모양이다. 오히려 위험도는 반대였다(Find는 스크롤백을 rewrap하는 write 지점,
cell_colors는 read). 실제 이유는 ⑴ ROI(42줄 대부분이 struct 리터럴) ⑵
io-render-threading.md **P4-3**이 F·G·I·J를 투영 tick의 단일 lock 스코프로 수렴시키려
한다는 것이다. 같은 이유로 이미 뺀 `collectFindViewSpans`(E)가 그 전제를 흔들므로,
그 문서 §12.2 행 E와 P4-3에 새 함수 경계와 "그냥 감싸면 비재진입 panic"을 등재했다.
`app-session-decomposition.md` §6에도 허브 추출 완료를 기록했다(수치·범위·검증 결과).

**게이트를 보강한다.** 앞선 세 커밋은 `test-macos-app-host-abi`와 `check-boundaries`만
보고했는데, 하필 손댄 곳이 프레임 루프와 포인터 제스처 라우팅이다. 이번에
`mise run check`(전체 `zig build test` 포함) · `zig build macos-app-build` ·
`mise run macos-app-smoke`를 전부 돌렸고 모두 통과했다.

digest는 갱신했다(`usize` 타입 변경이 토큰을 바꾼다 — 주석·doc은 토큰이 아니다).

+++ src/platform/macos/app_session.zig
@@ -22835,8 +22835,9 @@ pub const AppSession = struct {
         self.last_summary.last_event_kind = @intFromEnum(EventKind.key_down);
     }
 
-    /// 앱(chrome 오버레이·rename·주소창·사이드바 검색·도크·파일 트리 등)이 키를 소비했다.
-    /// PTY로 내려보내지 않고 여기서 끝낸다.
+    /// PTY write 없이 여기서 종결되는 모든 경로 — chrome 오버레이·rename·주소창·사이드바 검색·도크·
+    /// 파일 트리, 그리고 스크롤백 페이지 스크롤처럼 **터미널 자신이** 소비하는 경우까지 포함한다.
+    /// 열거가 아니라 이 판정 기준으로 읽어야 새 분기가 늘어도 낡지 않는다.
     fn keyConsumedByApp(self: *AppSession) FrameSummary {
         self.total_app_key_events += 1;
         self.settleKeyEventSummary();
@@ -23911,9 +23912,6 @@ pub const AppSession = struct {
         return layout_math.pxToCell(self.cell_width_px, self.cell_height_px, surface.core.size.cols, surface.core.size.rows, rect, x_px, y_px);
     }
 
-    /// 마우스 선택. kind 1=down(선택 시작), 2=drag(확장), 3=up(확정 — 드래그 선택인데 이동이
-    /// 없었으면 클릭으로 보고 해제), 4=더블클릭(단어 선택), 5=트리플클릭(논리 줄 선택). 좌표는
-    /// backing 픽셀 — 셀 변환은 권위 있는 cell 메트릭을 가진 여기서 한다.
     /// 진행 중인 포인터 제스처(사이드바 탭·그룹, 터미널 탭, pane, 각종 divider, 스크롤바 드래그)가
     /// 이 이벤트를 캡처하는지 판정하고, 캡처하면 처리까지 마친다. 반환 true = 이벤트 소비됨.
     ///
@@ -23931,7 +23929,7 @@ pub const AppSession = struct {
         // 반대로 **새 제스처를 시작하는 down은 오버레이가 먼저 본다** — 오버레이 위 클릭이 그 아래
         // 터미널·탭으로 새면 안 되기 때문이다. 그래서 이 구간은 kind 2/3만 잡고 kind 1은 아래로 흘린다.
         // 사이드바 탭 드래그가 진행 중이면 drag(2)/up(3)을 캡처한다(x가 사이드바 밖으로 나가도) — 새
-        // down(1)은 아래 일반 처리로 흘려 드래그를 새로 시작한다. drag는 타겟 슬롯으로 live 재정렬한다.
+        // down(1)은 호출자(`mouse`)의 일반 라우팅으로 흘려 드래그를 새로 시작한다. drag는 타겟 슬롯으로 live 재정렬한다.
         if (self.pointerGestureIs(.sidebar_tab) and (kind == 2 or kind == 3)) {
             const drag = &self.pointer_gesture_owner.sidebar_tab;
             if (kind == 2 and drag.index < self.tabs.items.len) {
@@ -24011,7 +24009,7 @@ pub const AppSession = struct {
             return true;
         }
         // Term 탭 드래그가 진행 중이면 drag(2)/up(3)을 캡처한다 — drag는 소스 pane 바 안에서 x로 타겟 탭을 잡아
-        // live 재정렬(PR-E1: pane 내). up이 끝낸다. 새 down(1)은 아래 일반 처리로 흘려 새 드래그를 시작한다.
+        // live 재정렬(PR-E1: pane 내). up이 끝낸다. 새 down(1)은 호출자(`mouse`)의 일반 라우팅으로 흘려 새 드래그를 시작한다.
         if (self.pointerGestureIs(.terminal_tab) and (kind == 2 or kind == 3)) {
             if (kind == 2) {
                 self.dragTabTo(x_px); // pane 내 live 재정렬(PR-E1)
@@ -24035,7 +24033,7 @@ pub const AppSession = struct {
         }
         // pane(분할 영역) 통째 드래그가 진행 중이면 drag(2)/up(3)을 캡처한다 — drag는 커서 좌표만 갱신(floating
         // 미리보기가 따라가게), up이 사이드바면 새 워크스페이스 분리/합치기(dropPaneAt). 사이드바 밖이면 no-op.
-        // 새 down(1)은 아래 일반 처리로 흘려 새 제스처를 시작한다.
+        // 새 down(1)은 호출자(`mouse`)의 일반 라우팅으로 흘려 새 제스처를 시작한다.
         if (self.pointerGestureIs(.pane) and (kind == 2 or kind == 3)) {
             const drag = &self.pointer_gesture_owner.pane;
             if (kind == 2) {
@@ -24110,6 +24108,9 @@ pub const AppSession = struct {
         return false;
     }
 
+    /// 마우스 선택. kind 1=down(선택 시작), 2=drag(확장), 3=up(확정 — 드래그 선택인데 이동이
+    /// 없었으면 클릭으로 보고 해제), 4=더블클릭(단어 선택), 5=트리플클릭(논리 줄 선택). 좌표는
+    /// backing 픽셀 — 셀 변환은 권위 있는 cell 메트릭을 가진 여기서 한다.
     pub fn mouse(self: *AppSession, kind: i32, x_px: f64, y_px: f64, button: i32, mods: i32) void {
         if (!self.surface_initialized) return;
         // 상태바 위 클릭은 **삼킨다**(S3가 항목을 올리기 전까지 눌러도 아무 일도 없는 게 맞다). 안 막으면 아래
@@ -30347,12 +30348,12 @@ pub const AppSession = struct {
     }
 
     /// 지난 프레임 동안 포인터가 남긴 것을 이번 tick에 정확히 한 번씩 적용하고, 살아 있는 탭 드래그가
-    /// 아직 유효한지 확인한다. 프레임 계측보다 **앞**에 둔다 — geometry가 확정된 상태로 아래 build와
-    /// 투영이 돌아야 한다.
+    /// 아직 유효한지 확인한다. 프레임 계측(`ft_start`)보다 **앞**에 둔다 — geometry가 확정된 상태로
+    /// 호출부(`tick`)의 build와 투영이 돌아야 한다.
     fn settleDeferredPointerInput(self: *AppSession) void {
         // 이 tick의 divider resize와 scrollbar scroll을 여기서 한 번씩 적용한다 — move가 몇 번 왔든 최종
         // 좌표 하나이고, clamp 결과가 직전과 같으면 effect가 없다(계약 §4.3). frame 계측이 시작되기 전에
-        // 두어 geometry가 확정된 상태로 아래 build/투영이 돌게 한다.
+        // 두어 geometry가 확정된 상태로 호출부(`tick`)의 build/투영이 돌게 한다.
         //
         // **둘 다 여기 있어야 한다.** scrollbar 쪽이 빠져 있어서 thumb을 끄는 동안에는 좌표만 쌓이고
         // 손을 뗄 때(`dropped`) 한 번에 적용됐다 — 같은 coalescing을 쓰는 divider는 매 tick 반영되는데
@@ -30378,10 +30379,16 @@ pub const AppSession = struct {
         if (self.pointerGestureIs(.terminal_tab) and self.anyModalOverlayOpen()) self.cancelPointerGesture();
     }
 
-    /// 프레임 계측이 시작되기 전 단계 — 저자가 `ft_pre` 마크에 "pre(housekeeping)"이라 이름 붙인 구간이다.
+    /// 프레임 계측의 첫 구간(`ft_start`→`ft_pre`) — 저자가 그 마크에 "pre(housekeeping)"이라 이름 붙인
+    /// 구간 그대로다. 계측 **전**에 도는 것은 `settleDeferredPointerInput` 쪽이므로 이 함수를 `ft_start`
+    /// 위로 올리면 pre 항목이 항상 0으로 붕괴한다.
+    ///
     /// 백그라운드 worker가 남긴 결과를 메인 스레드 상태로 옮기고(아카이브·파일 트리·git·업로드·업데이트),
-    /// 렌더 전에 성립해야 하는 불변식을 정리한다. 여기서 하는 일은 전부 큐에서 꺼내 적용하는 것뿐이고
-    /// 파일시스템·네트워크 I/O는 worker가 소유한다 — 이 순서가 무너지면 tick이 blocking I/O를 물게 된다.
+    /// 렌더 전에 성립해야 하는 불변식을 정리한다. **대부분은** worker 결과 큐를 꺼내 적용하는 것이지만
+    /// 예외가 둘이다 — `pollAgentKinds`는 ≈0.5s throttle로 foreground process 열거 syscall을 직접 하고
+    /// 그 안에서 surface core 락을 잡으며, `requestBranchMenu`(MARU_FORCE_BRANCH_MENU 경로)는 `.git`
+    /// 탐색과 git 실행 파일 탐색으로 `access(2)`를 친다. 새 항목을 여기 추가할 때는 이 둘 외에
+    /// blocking I/O를 늘리지 않는다 — 늘어나면 tick이 그만큼 blocking I/O를 물게 된다.
     fn runFramePreHousekeeping(self: *AppSession) void {
         if (!self.update_started) { // 첫 tick에서 인앱 새 버전 안내 백그라운드 체크를 1회 띄운다(config로 끔)
             self.update_started = true;
@@ -30464,12 +30471,16 @@ pub const AppSession = struct {
     }
 
     /// Find(⌘F) 하이라이트를 이 프레임 기준으로 다시 계산한다. 뷰포트에 보이는 매치를
-    /// `find_view_spans`에 채우고 현재 매치 span을 돌려준다 — 현재 매치만 별도 강조색으로 그리기 때문에
-    /// 나머지 span과 갈라서 반환한다.
+    /// `find_view_spans`에 채우고(Find 패널이 열려 있을 때만 — ⌘G 닫힘-네비 중에는 현재 매치 span만
+    /// 돌려준다) 현재 매치 span을 반환한다. 현재 매치만 별도 강조색으로 그리므로 나머지와 갈라서 준다.
     ///
-    /// `output_events`는 이 tick에 새 출력이 들어왔는지다. 출력이 있으면 매치의 절대 좌표가 스크롤백
-    /// eviction으로 어긋날 수 있어 재검색하고, 없으면 이전 결과를 그대로 클립만 한다.
-    fn collectFindViewSpans(self: *AppSession, output_events: u64) ?terminal.SelectionSpan {
+    /// `output_events`는 이 tick에 새 출력이 들어왔는지다. 출력이 있거나 원격 검색 결과가
+    /// dirty면(`remote_find_dirty`) 재검색하고, 그 외에는 이전 결과를 클립만 한다 — 출력이 들어오면
+    /// 매치의 절대 좌표가 스크롤백 eviction으로 어긋날 수 있기 때문이다.
+    ///
+    /// 활성 surface가 로컬 코어면 `lockCore`를 잡고 반환 전에 반드시 푼다. 이미 core 락을 쥔 자리에서
+    /// 부르면 `core_mutex`가 비재진입이라 `core_owner.zig`에서 즉시 panic한다(현재 호출자는 tick 하나).
+    fn collectFindViewSpans(self: *AppSession, output_events: usize) ?terminal.SelectionSpan {
         // Find 열린 채(또는 ⌘G 닫힘-네비 중) 새 출력이 들어오면 매치 절대 좌표가 어긋날 수 있다(스크롤백
         // eviction). 재검색해 하이라이트를 최신으로 유지하되 현재 인덱스만 clamp하고 스크롤은 하지 않는다.
         const find_active = self.chrome_host.find.open or self.find_nav;
@@ -30484,7 +30495,7 @@ pub const AppSession = struct {
             if (is_macos and fa_surface.remote != null) {
                 // §6c host-backed: 검색은 host가(콘텐츠·스크롤백 소유 — placeholder는 빈 셀이라 못 함). output(스크롤도
                 // delta라 포함)/query 변화 시만 host에 재검색(refreshRemoteFind)해 뷰포트 매치 span을 remote_find_spans에
-                // 캐시하고, tick은 그 캐시를 find_view_spans에 복사만 한다(매 tick RPC 회피). 현재 매치 강조색·네비는 #6c-2.
+                // 캐시하고, 이 함수는 그 캐시를 find_view_spans에 복사만 한다(매 tick RPC 회피). 현재 매치 강조색·네비는 #6c-2.
                 if (find_active) {
                     if (self.remote_find_dirty or output_events > 0) {
                         self.refreshRemoteFind(fa_surface);
